additional emscripten changes

+++ src/screen.c
@@ -115,8 +115,8 @@ void SDL_UpdateRects(SDL_Surface *screen, int numrects, SDL_Rect *rects)
 		SDL_UpdateTexture(sdlTexture, NULL, screen->pixels, screen->pitch);
 		/* Need to clear the renderer context for certain accelerated cards */
 		if (!bIsSoftwareRenderer)
-			SDL_RenderClear(sdlRenderer);
-		SDL_RenderCopy(sdlRenderer, sdlTexture, NULL, NULL);
+			(int)SDL_RenderClear(sdlRenderer);
+		(int)SDL_RenderCopy(sdlRenderer, sdlTexture, NULL, NULL);
 		SDL_RenderPresent(sdlRenderer);
 	}
 	else
@@ -511,13 +511,13 @@ static bool Screen_SetSDLVideoSize(int width, int height, int bitdepth, bool bFo
 		}
 
 		if (bInFullScreen)
-			SDL_RenderSetLogicalSize(sdlRenderer, width, height);
+			(int)SDL_RenderSetLogicalSize(sdlRenderer, width, height);
 		else
-			SDL_RenderSetScale(sdlRenderer, scale, scale);
+			(int)SDL_RenderSetScale(sdlRenderer, scale, scale);
 
 		/* Force to black to stop side bar artifacts on 16:9 monitors. */
-		SDL_SetRenderDrawColor(sdlRenderer, 0, 0, 0, 255);
-		SDL_RenderClear(sdlRenderer);
+		(int)SDL_SetRenderDrawColor(sdlRenderer, 0, 0, 0, 255);
+		(int)SDL_RenderClear(sdlRenderer);
 		SDL_RenderPresent(sdlRenderer);
 
 		SDL_GetRendererInfo(sdlRenderer, &sRenderInfo);
